+++ feature-hbti/src/main/java/com/hmoa/feature_hbti/screen/HbtiSurveyLoadingScreen.kt
@@ -17,7 +17,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import com.hmoa.component.TopBar
+import com.hmoa.core_designsystem.component.TopBar
 import com.hmoa.core_designsystem.component.HmoaLogoLoading
 import com.hmoa.core_designsystem.theme.pretendard
 import com.hmoa.feature_hbti.viewmodel.HbtiSurveyLoadingViewmodel
