+++ feature-hpedia/src/main/java/com/hmoa/feature_hpedia/Screen/HpediaSearchScreen.kt
@@ -27,7 +27,7 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.paging.compose.LazyPagingItems
 import androidx.paging.compose.collectAsLazyPagingItems
-import com.hmoa.component.TopBar
+import com.hmoa.core_designsystem.component.TopBar
 import com.hmoa.core_designsystem.R
 import com.hmoa.core_designsystem.component.SearchTopBar
 import com.hmoa.core_model.response.NoteDefaultResponseDto
