+++ feature-userInfo/src/main/java/com/hmoa/feature_userinfo/Screen/MyPage.kt
@@ -45,9 +45,9 @@ import androidx.compose.ui.viewinterop.AndroidView
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.google.android.gms.oss.licenses.OssLicensesMenuActivity
-import com.hmoa.component.TopBar
 import com.hmoa.core_common.ErrorUiState
 import com.hmoa.core_designsystem.R
+import com.hmoa.core_designsystem.component.TopBar
 import com.hmoa.core_designsystem.component.AppLoadingScreen
 import com.hmoa.core_designsystem.component.CircleImageView
 import com.hmoa.core_designsystem.component.ErrorUiSetView
@@ -55,7 +55,6 @@ import com.hmoa.core_designsystem.component.OnAndOffBtn
 import com.hmoa.core_designsystem.theme.CustomColor
 import com.hmoa.core_domain.entity.data.ColumnData
 import com.hmoa.feature_userinfo.BuildConfig
-import com.hmoa.feature_userinfo.Screen.NoAuthMyPage
 import com.hmoa.feature_userinfo.viewModel.MyPageViewModel
 import com.hmoa.feature_userinfo.viewModel.UserInfoUiState
 import com.kakao.sdk.talk.TalkApiClient
