+++ feature-userInfo/src/main/java/com/hmoa/feature_userinfo/Screen/NoDataPage.kt
@@ -1,4 +1,4 @@
-package com.hmoa.feature_userinfo
+package com.hmoa.feature_userinfo.screen
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
